ft-#76: move useSignDefaultFormValues to file
